refactor(ts-sdk): reduce custom-filters type tests to pure compile-time assertions
The vitest wrapper around these assertions could never fail at runtime —
the real gate is `tsc --noEmit`, where each @ts-expect-error directive
guards a line that only errors while `as const` narrowing is in effect
(removing the narrowing makes the directive unused, ts2578, failing the
type check). Drop the runtime test scaffolding and the widening-demo
block (which compiled clean by design and asserted nothing), and rename
.spec.ts -> .ts so vitest no longer collects the file. The as-const trap
remains documented in the extensions README and the example's comment
block.
